Arrange app.rb file
