+++ data/games.json
@@ -1,1 +1,1 @@
-[{"multiplayer":"multiplayer","publish_date":"2/18/2022","last_played_date":"02/20/2022"},{"multiplayer":"dog","publish_date":"2022-12-04","last_played_date":"2022-11-23"},{"multiplayer":"Faden","publish_date":"2022-02-23","last_played_date":"2022-01-22"},{"multiplayer":"doff","publish_date":"2021-11-03","last_played_date":"1995-02-22"},{"multiplayer":"fooor","publish_date":"2004-12-20","last_played_date":"2018-11-19"},{"multiplayer":"gkgkg","publish_date":"2003-03-30","last_played_date":"2021-03-12"},{"multiplayer":"yooor","publish_date":"2008-10-11","last_played_date":"2019-03-22"},{"multiplayer":"multi","publish_date":"2022-12-11","last_played_date":"2025-11-01"},{"multiplayer":"asd","publish_date":"11-3-2022","last_played_date":"11-2-2022"},{"multiplayer":"dsa","publish_date":"30-3-2022","last_played_date":"21-2-2022"},{"multiplayer":"zxc","publish_date":"22-5-2024","last_played_date":"2-1-2023"},{"multiplayer":"qas","publish_date":"1-3-2022","last_played_date":"1-1-2022"}]
+[{"multiplayer":"multiplayer","publish_date":"2/18/2022","last_played_date":"02/20/2022"},{"multiplayer":"dog","publish_date":"2022-12-04","last_played_date":"2022-11-23"},{"multiplayer":"Faden","publish_date":"2022-02-23","last_played_date":"2022-01-22"},{"multiplayer":"doff","publish_date":"2021-11-03","last_played_date":"1995-02-22"},{"multiplayer":"fooor","publish_date":"2004-12-20","last_played_date":"2018-11-19"},{"multiplayer":"gkgkg","publish_date":"2003-03-30","last_played_date":"2021-03-12"},{"multiplayer":"yooor","publish_date":"2008-10-11","last_played_date":"2019-03-22"},{"multiplayer":"multi","publish_date":"2022-12-11","last_played_date":"2025-11-01"},{"multiplayer":"asd","publish_date":"11-3-2022","last_played_date":"11-2-2022"},{"multiplayer":"dsa","publish_date":"30-3-2022","last_played_date":"21-2-2022"},{"multiplayer":"zxc","publish_date":"22-5-2024","last_played_date":"2-1-2023"},{"multiplayer":"qas","publish_date":"1-3-2022","last_played_date":"1-1-2022"},{"multiplayer":"AAAA","publish_date":"1999-03-08","last_played_date":"1999-03-08"}]
